requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pika>=1.3.2
PyYAML>=6.0
numpy>=1.25.0
tqdm>=4.66.0
matplotlib>=3.8.0